Challenge #2-Know_Your_Fan_Parte_1/requirements.txt
# Framework Web
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Banco de Dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# AWS
boto3==1.29.3
awscli==1.32.3

# Machine Learning
tensorflow==2.14.0
torch==2.1.1
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2

# Análise de Dados
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0

# APIs e Web Scraping
tweepy==4.14.0
instagram-private-api==1.6.0
google-api-python-client==2.108.0
beautifulsoup4==4.12.2
requests==2.31.0

# Utilitários
python-dotenv==1.0.0
pydantic==2.5.2
pytest==7.4.3
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Documentação
sphinx==7.2.6
mkdocs==1.5.3
mkdocs-material==9.4.14 